[gameui/matchmaking] Adjust option index for x86-64

+++ gameui/matchmaking/dialogmenu.h
@@ -158,12 +158,12 @@ public:
 	void	ApplySchemeSettings( vgui::IScheme *pScheme ) override;
 	void	SetFocus( const bool bActive ) override;
 
-	void			SetOptionFocus( unsigned int idx );
+	void			SetOptionFocus( intp idx );
 	void			SetOptionFocusNext();
 	void			SetOptionFocusPrev();
 
 	void					AddOption( const char *pLabelText, const sessionProperty_t &option );
-	int						GetActiveOptionIndex();
+	intp					GetActiveOptionIndex();
 	const sessionProperty_t &GetActiveOption();
 
 	void			DeleteAllOptions()
@@ -173,7 +173,7 @@ public:
 		m_nActiveOption = m_Options.InvalidIndex();
 	}
 private:
-	int				m_nActiveOption;
+	intp			m_nActiveOption;
 	int				m_nOptionsXPos;
 	int				m_nOptionsMinWide;
 	int				m_nOptionsLeftMargin;
@@ -310,7 +310,7 @@ public:
 	int					GetItemCount();
 	int					GetActiveItemIndex();
 	int					GetActiveColumnIndex();
-	int					GetActiveOptionIndex( int idx );
+	intp				GetActiveOptionIndex( int idx );
 	int					GetVisibleItemCount();
 	int					GetVisibleColumnCount();
 	int					GetFirstUnlockedColumnIndex();
